feat: add Kampfkarren/selene
[Kampfkarren/selene](https://github.com/Kampfkarren/selene): A blazing-fast modern Lua linter written in Rust

```console
$ aqua g -i Kampfkarren/selene
```

Reviewers aren't necessarily familiar with this package, so please describe how to confirm if this package works well.
Please confirm if this package works well yourself as much as possible.

Command and output

```console
$ echo 'print(123)' | selene -
Results:
0 errors
0 warnings
0 parse errors
$ echo 'a = 123' | selene -
warning[unscoped_variables]: `a` is not declared locally, and will be available in every scope
  ┌─ -:1:1
  │
1 │ a = 123
  │ ^

warning[unused_variable]: a is defined, but never used
  ┌─ -:1:1
  │
1 │ a = 123
  │ ^

Results:
0 errors
2 warnings
0 parse errors

```

If files such as configuration file are needed, please share them.

None

Reference

- https://kampfkarren.github.io/selene/
- https://github.com/Kampfkarren/selene/releases/tag/0.24.0